Edited wlui and added tests.
Added 2 simple functions to wlui module., and 2 tests to worklog3_test module.

+++ wlui.py
@@ -22,6 +22,16 @@ def clear_screen():
     os.system('cls' if os.name == 'nt' else 'clear')
 
 
+def invalid_input(wanted):
+    """Warn user about incorrect input."""
+    input("Invalid {}.  Please try again.".format(wanted))
+    return
+
+
+def press_enter():
+    input("Press enter to return.")
+    return
+
 def add_entry():
     """Add new entry to work log."""
     clear_screen()
@@ -29,16 +39,16 @@ def add_entry():
     while name_adding:
         add_name = input('Please enter full name.\n> ')
         if CardCatalog().name_check(add_name):
-            name_adding -= 1
+            break
         else:
-            input('Please enter first and last name for employee.')
+            invalid_input('first and last name string')
             continue
 
     task_adding = 1
     while task_adding:
         add_task = input('What task did you work on?\n> ')
         if isinstance(add_task, str):
-            task_adding -= 1
+            break
         else:
             input('Task name must be a string.')
             continue
@@ -47,9 +57,9 @@ def add_entry():
     while minutes_adding:
         add_minutes = input('How many minutes were spent?\n> ')
         if CardCatalog().minute_check(add_minutes):
-            minutes_adding -= 1
+            break
         else:
-            input('Invalid integer. Press enter to try again.')
+            invalid_input('integer')
             continue
 
     add_notes = ''
@@ -87,7 +97,7 @@ def edit_entry(entry):
         new_minutes = input('Edit integer of minutes, or enter to skip.\n> ')
         if new_minutes:
             if CardCatalog().minute_check(new_minutes):
-                nm_check -= 1
+                break
             else:
                 continue
 
@@ -107,9 +117,9 @@ def edit_entry(entry):
                     new_date,
                     "%m/%d/%Y"
                 )
-                nd_check -= 1
+                break
             except ValueError:
-                input('Invalid date. Press enter to retry.')
+                invalid_input('date string')
                 continue
 
         new_notes = ''
@@ -147,10 +157,11 @@ def edit_entry(entry):
 
             saver = CardCatalog().save_edits(updater)
             if saver:
-                input("Saved succesfully! Hit enter to continue.")
+                print("Saved succesfully!")
+                press_enter()
             else:
-                input("Hit enter to return to previous menu.")
-            entry_adding -= 1
+                press_enter()
+            return
 
 
 def by_staff():
@@ -186,20 +197,21 @@ Please choose:
         timeloop = 1
         while timeloop:
             clear_screen()
+            # My project reviewer reported 'an error searching with "03/03/3333-05/05/5555". The same thing happens if I try to pick a date from the list with a bad input.' By the time I tried to recrete said error, I'd already implemented the d_range_check function to catch improper date range inputs.  So, I think I fixed it?
             target_range = input('''
 Please enter date range in following format:
 "MM/DD/YYYY-MM/DD/YYYY"
 Range will find from midnight of first date, to midnight of the last date.
 So, if you want to search New Year's Day 2018, type "01/01/2018-01/02/2018"
 > ''')
-            targets = target_range.split(sep='-')
-            dateobj = CardCatalog().date_check(targets[0])
-            dateobj1 = CardCatalog().date_check(targets[1])
-            if dateobj & dateobj1:
+            if CardCatalog().d_range_check(target_range):
+                targets = target_range.split(sep='-')
+                dateobj = CardCatalog().date_check(targets[0])
+                dateobj1 = CardCatalog().date_check(targets[1])
                 view_entries(bycat='date', target=dateobj, datelast=dateobj1)
                 return
             else:
-                input('Invalid date string.  Press enter to retry.')
+                invalid_input('date range string')
     else:
         datelog = CardCatalog().generate_datelog()
         d_num = 0
@@ -208,10 +220,16 @@ So, if you want to search New Year's Day 2018, type "01/01/2018-01/02/2018"
                                   fenix.strftime("%A %B %d, %Y")))
             print('-' * 50)
             d_num += 1
-        snackfruit = input("Please type a number to select a date. \n> ")
-        view_entries(bycat='date', target=datelog[int(snackfruit)])
-        clear_screen()
-        return
+        listloop = 1
+        while listloop:
+            snackfruit = input("Please type a number to select a date. \n> ")
+            # Check that number entered exists in list.
+            if int(snackfruit) >= 0 and int(snackfruit) < len(datelog):
+                view_entries(bycat='date', target=datelog[int(snackfruit)])
+                clear_screen()
+                return
+            else:
+                invalid_input('date string')
 
 
 def by_minutes():
@@ -223,7 +241,7 @@ def by_minutes():
             view_entries(bycat='minutes', target=minnum)
             return
         else:
-            input('Invalid input.  Press enter to continue.')
+            invalid_input('integer')
             clear_screen()
 
 
@@ -243,48 +261,51 @@ def view_entries(bycat=None, target=None, datelast=None):
     else:
         viewer = CardCatalog().load_entries(bycat=bycat,
                                             target=target)
-    view = len(viewer)
-    view_num = 0
-    while view_num < view:
-        clear_screen()
-        vts = viewer[view_num]['timestamp'].strftime('%A %B %d')
-        print(vts)
-        print('='*len(vts))
-        print("Recorded by: " + viewer[view_num]['user_name'])
-        print("Task: " + viewer[view_num]['task_name'])
-        print("Minutes: " + str(viewer[view_num]['task_minutes']))
-        print("Additional Notes: " + viewer[view_num]['task_notes'])
-        print('\n\n' + '=' * len(vts))
-        print('N) Next entry')
-        print('p) Previous entry')
-        print('e) Edit entry')
-        print('d) Delete entry')
-        print('q) Return to menu')
-
-        next_action = input('Action: \n[N/p/e/d/q]: ').lower().strip()
-        if next_action == 'q':
-            break
-            # Break exits view loop, returns to main menu.
-        elif next_action == 'e':
-            edit_entry(viewer[view_num])
-            break
-        elif next_action == 'd':
-            if input('Delete entry. Are you sure? y/N ').lower() == 'y':
-                CardCatalog().delete_entry(viewer[view_num])
+    if viewer:
+        view = len(viewer)
+        view_num = 0
+        while view_num < view:
+            clear_screen()
+            vts = viewer[view_num]['timestamp'].strftime('%A %B %d, %Y')
+            print(vts)
+            print('='*len(vts))
+            print("Recorded by: " + viewer[view_num]['user_name'])
+            print("Task: " + viewer[view_num]['task_name'])
+            print("Minutes: " + str(viewer[view_num]['task_minutes']))
+            print("Additional Notes: " + viewer[view_num]['task_notes'])
+            print('\n\n' + '=' * len(vts))
+            print('N) Next entry')
+            print('p) Previous entry')
+            print('e) Edit entry')
+            print('d) Delete entry')
+            print('q) Return to menu')
+    
+            next_action = input('Action: \n[N/p/e/d/q]: ').lower().strip()
+            if next_action == 'q':
                 break
+                # Break exits view loop, returns to main menu.
+            elif next_action == 'e':
+                edit_entry(viewer[view_num])
+                break
+            elif next_action == 'd':
+                if input('Delete entry. Are you sure? y/N ').lower() == 'y':
+                    CardCatalog().delete_entry(viewer[view_num])
+                    break
+                else:
+                    view_num += 0
+            elif next_action == 'p':
+                if view_num > 0:
+                    view_num -= 1
+                else:
+                    input("Can't go further than that. (Enter to continue.)")
             else:
-                view_num += 0
-        elif next_action == 'p':
-            if view_num > 0:
-                view_num -= 1
-            else:
-                input("Can't go further than that. (Enter to continue.)")
-        else:
-            if view_num < view - 1:
-                view_num += 1
-            else:
-                input("That's all we could find. (Enter to continue.)")
-
+                if view_num < view - 1:
+                    view_num += 1
+                else:
+                    input("That's all we could find. (Enter to continue.)")
+    else:
+        press_enter()
+        return
 
 def search_entries():
     """Search existing entries."""
